feat: enable dynamic plugin deployment
This change introduces the changes needed to deploy these regular
backstage plugins to OCP running in RHDH.  A workaround was added to the
backend plugin's package.json for RHIDP-2945 to lock the version of
@backstage/backend-plugin-api.  The backend code had to be changed
slightly to avoid importing the root router from
@backstage/backend-defaults.  Both frontend and backend plugins have had
export-dynamic scripts added to the package.json.

A "deploy" directory has been added to stage the .tgz files.  3 scripts
have been added to aid deploying the files to openshift, one to stage
the binary .tgz files.  One to initially create the plugin-registry
build and app, and a third to update the plugin-registry image as
needed.

Finally, a default configuration snippet has been added to the frontend
plugin with the required frontend configuration for this plugin to
appear in the RHDH UI.