chore: update package versions and text
This change bumps some package versions of the plugins and updates the
@janus-idp/cli version to fix a bug.  This change also updates the
readme text accordingly, removing mention of the workaround that was
needed on previous CLI versions.  This change also adds a little script
and mention in the appendix on a quick way to run a local developer hub
instance to try out the plugins.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>